+++ src/IbanNet/Registry/IbanRegistry.cs
@@ -40,8 +40,11 @@ public class IbanRegistry : IIbanRegistry
     /// <inheritdoc />
     public int Count => Dictionary.Count;
 
-    /// <inheritdoc />
-    public IList<IIbanRegistryProvider> Providers { get; internal init; } = new List<IIbanRegistryProvider>();
+    /// <summary>
+    /// Gets the registry providers.
+    /// </summary>
+    /// <remarks>Note: after using any of the registry members (iow. getting country information), you can no longer modify this list.</remarks>
+    public IList<IIbanRegistryProvider> Providers { get; internal set; } = new List<IIbanRegistryProvider>();
 
     /// <inheritdoc />
     // ReSharper disable once InconsistentNaming
@@ -66,12 +69,21 @@ public class IbanRegistry : IIbanRegistry
 
             lock (_syncObject)
             {
-                return _dictionary ??= new ReadOnlyDictionary<string, IbanCountry>(Providers
-                    .SelectMany(p => p)
-                    // In case of duplicate country codes, select the first.
-                    .GroupBy(c => c.TwoLetterISORegionName)
-                    .ToDictionary(g => g.Key, g => g.First(), StringComparer.OrdinalIgnoreCase)
-                );
+                var readOnlyProviders = new ReadOnlyCollection<IIbanRegistryProvider>(Providers.ToArray());
+                try
+                {
+                    return _dictionary ??= new ReadOnlyDictionary<string, IbanCountry>(readOnlyProviders
+                        .SelectMany(p => p)
+                        // In case of duplicate country codes, select the first.
+                        .GroupBy(c => c.TwoLetterISORegionName)
+                        .ToDictionary(g => g.Key, g => g.First(), StringComparer.OrdinalIgnoreCase)
+                    );
+                }
+                finally
+                {
+                    // Freeze providers, so that the collection can no longer be modified after hydrating the registry.
+                    Providers = readOnlyProviders;
+                }
             }
         }
     }